Extract refreshFolderList command to CommandRefreshFolderList
