+++ backend/demo/src/main/java/app/k9mail/backend/demo/CommandRefreshFolderList.kt
@@ -1,0 +1,29 @@
+package app.k9mail.backend.demo
+
+import com.fsck.k9.backend.api.BackendStorage
+import com.fsck.k9.backend.api.FolderInfo
+import com.fsck.k9.backend.api.updateFolders
+
+internal class CommandRefreshFolderList(
+    private val backendStorage: BackendStorage,
+    private val demoStore: DemoStore,
+) {
+
+    fun refreshFolderList() {
+        val localFolderServerIds = backendStorage.getFolderServerIds().toSet()
+
+        backendStorage.updateFolders {
+            val remoteFolderServerIds = demoStore.getFolderIds()
+            val foldersServerIdsToCreate = remoteFolderServerIds - localFolderServerIds
+            val foldersToCreate = foldersServerIdsToCreate.mapNotNull { folderServerId ->
+                demoStore.getFolder(folderServerId)?.let { folderData ->
+                    FolderInfo(folderServerId, folderData.name, folderData.type)
+                }
+            }
+            createFolders(foldersToCreate)
+
+            val folderServerIdsToRemove = (localFolderServerIds - remoteFolderServerIds).toList()
+            deleteFolders(folderServerIdsToRemove)
+        }
+    }
+}
